Updates block_hasher to be a verifier.
It now takes in the data in a json file, and returns whether the block is
valid or not. The verify method is now exported from the verifier, so we
can run a test script to test it. Adds an invalid data block for testing,
and adds a test script which runs the verifier against both the valid and
invalid data.