Separates out the model from the view
The views now get their data from their js controllers, rather than having
hard coded values in the html. For now, the data model has been
implemented as a map in the controllers class. We use the name of the
block as the key. Updates the unit and e2e tests.