Replaces the carousel with one provided by the UI Bootstrap project
Adds a dependency on the UI Bootstrap project, which has been added to
the main app and to the Karma config file. The UI Bootstrap project
provides you with a <carousel> tag which you can use in the html. This
does make the code simpler to read. The buttons on the new carousel
control work as well. One small niggle is that the control flashes for a
second when we filter it. I've gone with this for now anyway.