Updates the block explorer app so it can query bitcoin explorer
To do this, we proxy the bitcoin explorer site to get around CORS header
issues. Because our app is running at the same origin as the proxy, we
are able to make requests to get block information.

Updates the package.json file to run our server instead of the default one
the seed project starts up.