Updates the data stored by our model
We now just store the hashes we have retrieved rather than storing all
the block data. This is because the amount of block data stored could get
very large if we were querying a large number of blocks, for example by
going back and forward through pages of 10. We currently query the
bitcoin service each time we choose a block by hash. We could improve
this by storing block data locally, and then loading it locally as needed.

We no longer use the carousel on the main page, as it makes an assumption
we know how many pages we have. In reality, we will use next and previous
buttons to navigate pages, and it is likely that going back will load
a new page of data. The next and previous buttons are wired up to our
controller, but the controller doesn't do anything with those commands yet.