Update our data model so it contains multiple pages of data
As we go back a page, we load in the previous 10 blocks into an array
which we prepend to our existing data. Each page of 10 hashes is stored in
an array, and the arrays are stored within our model. I've now added
implementations for the next and previous functions within our block list
controller.